Typo pointed out by Javier Goizueta fixed.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@4319 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/bigdecimal/lib/bigdecimal/util.rb
@@ -20,7 +20,7 @@
 #
 class Float < Numeric
   def to_d
-    BigFloat::new(selt.to_s)
+    BigDecimal::new(selt.to_s)
   end
 end
